execution/requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0

# Optional dependencies
playwright>=1.40.0  # For dynamic content scraping
openpyxl>=3.1.0  # For Excel file support

# Google API
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Environment management
python-dotenv>=1.0.0
